+++ bfvmm/include/hve/arch/x64/gdt.h
@@ -459,7 +459,9 @@ public:
         this->set_limit(index, limit);
     }
 
+#ifndef ENABLE_BUILD_TEST
 private:
+#endif
 
     /// @cond
