+++ bfvmm/src/vcpu/CMakeLists.txt
@@ -24,8 +24,13 @@ list(APPEND SOURCES
 if(${BUILD_TARGET_ARCH} STREQUAL "x86_64")
     list(APPEND SOURCES
         arch/intel_x64/vcpu.cpp
-        arch/intel_x64/vcpu_factory.cpp
     )
+
+    if(NOT ENABLE_MOCKING)
+        list(APPEND SOURCES
+            arch/intel_x64/vcpu_factory.cpp
+        )
+    endif()
 elseif(${BUILD_TARGET_ARCH} STREQUAL "aarch64")
     message(WARNING "Unimplemented")
 else()
