+++ bfvmm/tests/CMakeLists.txt
@@ -23,13 +23,12 @@ include(${SOURCE_CMAKE_DIR}/project.cmake)
 init_project()
 
 add_subdirectory(../src/debug ${CMAKE_CURRENT_BINARY_DIR}/src/debug)
-add_subdirectory(../src/entry ${CMAKE_CURRENT_BINARY_DIR}/src/entry)
 add_subdirectory(../src/hve ${CMAKE_CURRENT_BINARY_DIR}/src/hve)
 add_subdirectory(../src/memory_manager ${CMAKE_CURRENT_BINARY_DIR}/src/memory_manager)
 add_subdirectory(../src/vcpu ${CMAKE_CURRENT_BINARY_DIR}/src/vcpu)
 
 add_subdirectory(debug)
-# add_subdirectory(entry)
 add_subdirectory(hve)
 add_subdirectory(memory_manager)
-# add_subdirectory(vcpu)
+add_subdirectory(vcpu)
+add_subdirectory(support)
