+++ bfvmm/tests/hve/arch/intel_x64/vmcs/test_vmcs_check_controls.cpp
@@ -15,76 +15,16 @@
 // You should have received a copy of the GNU Lesser General Public
 // License along with this library; if not, write to the Free Software
 
-#include <catch/catch.hpp>
-#include <hippomocks.h>
-
-#include <intrinsics/x86/intel_x64.h>
-#include <intrinsics/x86/common_x64.h>
-#include <memory_manager/memory_manager_x64.h>
+#include <support/arch/intel_x64/test_support.h>
 
 #ifdef _HIPPOMOCKS__ENABLE_CFUNC_MOCKING_SUPPORT
-
-using namespace intel_x64;
-using namespace msrs;
-using namespace vmcs;
-using namespace check;
-
-struct cpuid_regs g_cpuid_regs;
-
-
-bool g_phys_to_virt_return_nullptr = false;
-uint64_t g_test_addr = 0U;
-uint64_t g_virt_apic_addr = 0U;
-uint8_t g_virt_apic_mem[0x81] = {0U};
-
-std::map<uint64_t, void *> g_mock_mem {
-    {
-        {g_virt_apic_addr, static_cast<void *>(&g_virt_apic_mem)},
-    }
-};
-
-uint32_t
-test_cpuid_eax(uint32_t val) noexcept
-{ return g_eax_cpuid[val]; }
-
-uint64_t
-test_read_msr(uint32_t addr) noexcept
-{ return g_msrs[addr]; }
-
-static bool
-test_vmread(uint64_t field, uint64_t *val) noexcept
-{
-    *val = g_vmcs_fields[field];
-    return true;
-}
-
-static bool
-test_vmwrite(uint64_t field, uint64_t val) noexcept
-{
-    g_vmcs_fields[field] = val;
-    return true;
-}
-
-static void
-setup_intrinsics(MockRepository &mocks, memory_manager_x64 *mm)
-{
-    mocks.OnCallFunc(memory_manager_x64::instance).Return(mm);
-    mocks.OnCall(mm, memory_manager_x64::physint_to_virtptr).Do(test_physint_to_virtptr);
-    mocks.OnCallFunc(_cpuid_eax).Do(test_cpuid_eax);
-    mocks.OnCallFunc(_read_msr).Do(test_read_msr);
-    mocks.OnCallFunc(_vmread).Do(test_vmread);
-    mocks.OnCallFunc(_vmwrite).Do(test_vmwrite);
-}
-
-static struct control_flow_path path;
 
 void
 test_vmcs_check(std::vector<struct control_flow_path> cfg, void(*func)())
 {
     for (auto p : cfg) {
         MockRepository mocks;
-        auto mm = mocks.Mock<memory_manager_x64>();
-        setup_intrinsics(mocks, mm);
+        setup_mm(mocks);
 
         p.setup();
 
@@ -101,11 +41,11 @@ static void
 setup_check_control_pin_based_ctls_reserved_properly_set_paths(std::vector<struct control_flow_path>
         &cfg)
 {
-    path.setup = [&] { g_msrs[ia32_vmx_true_pinbased_ctls::addr] = 0xffffffff00000000UL; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_pinbased_ctls::addr] = 0xffffffff00000000UL; };
     path.throws_exception = false;
     cfg.push_back(path);
 
-    path.setup = [&] { g_msrs[ia32_vmx_true_pinbased_ctls::addr] = 1; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_pinbased_ctls::addr] = 1; };
     path.throws_exception = true;
     cfg.push_back(path);
 
@@ -118,11 +58,11 @@ static void
 setup_check_control_proc_based_ctls_reserved_properly_set_paths(
     std::vector<struct control_flow_path> &cfg)
 {
-    path.setup = [&] { g_msrs[ia32_vmx_true_procbased_ctls::addr] = 0xffffffff00000000UL; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_procbased_ctls::addr] = 0xffffffff00000000UL; };
     path.throws_exception = false;
     cfg.push_back(path);
 
-    path.setup = [&] { g_msrs[ia32_vmx_true_procbased_ctls::addr] = 1; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_procbased_ctls::addr] = 1; };
     path.throws_exception = true;
     cfg.push_back(path);
 
@@ -135,19 +75,19 @@ static void
 setup_check_control_proc_based_ctls2_reserved_properly_set_paths(
     std::vector<struct control_flow_path> &cfg)
 {
-    path.setup = [&] { proc_ctl_disallow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask); };
+    path.setup = [&] { proc_ctl_disallow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask); };
     path.throws_exception = true;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        g_msrs[ia32_vmx_procbased_ctls2::addr] = 0xffffffff00000000UL;
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        g_msrs[intel_x64::msrs::ia32_vmx_procbased_ctls2::addr] = 0xffffffff00000000UL;
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        g_msrs[ia32_vmx_procbased_ctls2::addr] |= 1;
+        g_msrs[intel_x64::msrs::ia32_vmx_procbased_ctls2::addr] |= 1;
         secondary_processor_based_vm_execution_controls::set(0UL);
     };
     path.throws_exception = true;
@@ -156,7 +96,7 @@ setup_check_control_proc_based_ctls2_reserved_properly_set_paths(
     path.setup = [&] {
         using namespace secondary_processor_based_vm_execution_controls;
 
-        g_msrs[ia32_vmx_procbased_ctls2::addr] = 0xfffffffe00000000UL;
+        g_msrs[intel_x64::msrs::ia32_vmx_procbased_ctls2::addr] = 0xfffffffe00000000UL;
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
 
         // we use the _fields_ set() here rather than the controls enable()
@@ -187,7 +127,7 @@ static void
 setup_check_control_io_bitmap_address_bits_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::use_io_bitmaps::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::use_io_bitmaps::mask);
         primary_processor_based_vm_execution_controls::use_io_bitmaps::disable();
         g_eax_cpuid[x64::cpuid::addr_size::addr] = 48U;
     };
@@ -195,7 +135,7 @@ setup_check_control_io_bitmap_address_bits_paths(std::vector<struct control_flow
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::use_io_bitmaps::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::use_io_bitmaps::mask);
         primary_processor_based_vm_execution_controls::use_io_bitmaps::enable();
         address_of_io_bitmap_a::set(0x1U);
     };
@@ -226,14 +166,14 @@ static void
 setup_check_control_msr_bitmap_address_bits_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::use_msr_bitmap::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::use_msr_bitmap::mask);
         primary_processor_based_vm_execution_controls::use_msr_bitmap::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::use_msr_bitmap::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::use_msr_bitmap::mask);
         primary_processor_based_vm_execution_controls::use_msr_bitmap::enable();
         address_of_msr_bitmap::set(0x1U);
     };
@@ -301,13 +241,13 @@ setup_check_control_tpr_shadow_and_virtual_apic_paths(std::vector<struct control
     path.setup = [&] {
         proc_ctl2_allow0(virtualize_apic_accesses::mask);
         virtualize_apic_accesses::disable();
-        g_phys_to_virt_return_nullptr = true;
+        g_phys_to_virt_fails = true;
     };
     path.throws_exception = true;
     cfg.push_back(path);
 
     path.setup = [&] {
-        g_phys_to_virt_return_nullptr = false;
+        g_phys_to_virt_fails = false;
         g_test_addr = g_virt_apic_addr;
     };
     path.throws_exception = false;
@@ -368,15 +308,15 @@ static void
 setup_check_control_nmi_exiting_and_virtual_nmi_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        pin_ctl_allow1(ia32_vmx_true_pinbased_ctls::nmi_exiting::mask);
+        pin_ctl_allow1(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::nmi_exiting::mask);
         pin_based_vm_execution_controls::nmi_exiting::enable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        pin_ctl_allow0(ia32_vmx_true_pinbased_ctls::nmi_exiting::mask);
-        pin_ctl_allow1(ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
+        pin_ctl_allow0(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::nmi_exiting::mask);
+        pin_ctl_allow1(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
         pin_based_vm_execution_controls::nmi_exiting::disable();
         pin_based_vm_execution_controls::virtual_nmis::enable();
     };
@@ -384,7 +324,7 @@ setup_check_control_nmi_exiting_and_virtual_nmi_paths(std::vector<struct control
     cfg.push_back(path);
 
     path.setup = [&] {
-        pin_ctl_allow0(ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
+        pin_ctl_allow0(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
         pin_based_vm_execution_controls::virtual_nmis::disable();
     };
     path.throws_exception = false;
@@ -395,15 +335,15 @@ static void
 setup_check_control_virtual_nmi_and_nmi_window_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        pin_ctl_allow1(ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
+        pin_ctl_allow1(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
         pin_based_vm_execution_controls::virtual_nmis::enable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        pin_ctl_allow1(ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::nmi_window_exiting::mask);
+        pin_ctl_allow1(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::virtual_nmis::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::nmi_window_exiting::mask);
         pin_based_vm_execution_controls::virtual_nmis::disable();
         primary_processor_based_vm_execution_controls::nmi_window_exiting::enable();
     };
@@ -411,7 +351,7 @@ setup_check_control_virtual_nmi_and_nmi_window_paths(std::vector<struct control_
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::nmi_window_exiting::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::nmi_window_exiting::mask);
         primary_processor_based_vm_execution_controls::nmi_window_exiting::disable();
     };
     path.throws_exception = false;
@@ -422,14 +362,14 @@ static void
 setup_check_control_virtual_apic_address_bits_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::virtualize_apic_accesses::disable();
     };
@@ -463,15 +403,15 @@ setup_check_control_x2apic_mode_and_virtual_apic_access_paths(std::vector<struct
         &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::virtualize_x2apic_mode::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtualize_x2apic_mode::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::virtualize_x2apic_mode::disable();
     };
@@ -479,9 +419,9 @@ setup_check_control_x2apic_mode_and_virtual_apic_access_paths(std::vector<struct
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::virtualize_x2apic_mode::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::virtualize_apic_accesses::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtualize_x2apic_mode::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtualize_apic_accesses::mask);
         secondary_processor_based_vm_execution_controls::virtualize_x2apic_mode::enable();
         secondary_processor_based_vm_execution_controls::virtualize_apic_accesses::enable();
     };
@@ -489,8 +429,8 @@ setup_check_control_x2apic_mode_and_virtual_apic_access_paths(std::vector<struct
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::virtualize_x2apic_mode::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtualize_x2apic_mode::mask);
         secondary_processor_based_vm_execution_controls::virtualize_x2apic_mode::disable();
     };
     path.throws_exception = false;
@@ -502,15 +442,15 @@ setup_check_control_virtual_interrupt_and_external_interrupt_paths(
     std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::virtual_interrupt_delivery::disable();
     };
@@ -518,9 +458,9 @@ setup_check_control_virtual_interrupt_and_external_interrupt_paths(
     cfg.push_back(path);
 
     path.setup = [&] {
-        pin_ctl_allow0(ia32_vmx_true_pinbased_ctls::external_interrupt_exiting::mask);
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
+        pin_ctl_allow0(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::external_interrupt_exiting::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
         secondary_processor_based_vm_execution_controls::virtual_interrupt_delivery::enable();
         pin_based_vm_execution_controls::external_interrupt_exiting::disable();
     };
@@ -528,7 +468,7 @@ setup_check_control_virtual_interrupt_and_external_interrupt_paths(
     cfg.push_back(path);
 
     path.setup = [&] {
-        pin_ctl_allow1(ia32_vmx_true_pinbased_ctls::external_interrupt_exiting::mask);
+        pin_ctl_allow1(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::external_interrupt_exiting::mask);
         pin_based_vm_execution_controls::external_interrupt_exiting::enable();
     };
     path.throws_exception = false;
@@ -540,15 +480,15 @@ setup_check_control_process_posted_interrupt_checks_paths(std::vector<struct con
         &cfg)
 {
     path.setup = [&] {
-        pin_ctl_allow0(ia32_vmx_true_pinbased_ctls::process_posted_interrupts::mask);
+        pin_ctl_allow0(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::process_posted_interrupts::mask);
         pin_based_vm_execution_controls::process_posted_interrupts::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        pin_ctl_allow1(ia32_vmx_true_pinbased_ctls::process_posted_interrupts::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        pin_ctl_allow1(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::process_posted_interrupts::mask);
         pin_based_vm_execution_controls::process_posted_interrupts::enable();
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
@@ -556,8 +496,8 @@ setup_check_control_process_posted_interrupt_checks_paths(std::vector<struct con
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::virtual_interrupt_delivery::disable();
     };
@@ -565,9 +505,9 @@ setup_check_control_process_posted_interrupt_checks_paths(std::vector<struct con
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
-        exit_ctl_allow0(ia32_vmx_true_exit_ctls::acknowledge_interrupt_on_exit::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
+        exit_ctl_allow0(intel_x64::msrs::ia32_vmx_true_exit_ctls::acknowledge_interrupt_on_exit::mask);
         secondary_processor_based_vm_execution_controls::virtual_interrupt_delivery::enable();
         vm_exit_controls::acknowledge_interrupt_on_exit::disable();
     };
@@ -575,7 +515,7 @@ setup_check_control_process_posted_interrupt_checks_paths(std::vector<struct con
     cfg.push_back(path);
 
     path.setup = [&] {
-        exit_ctl_allow1(ia32_vmx_true_exit_ctls::acknowledge_interrupt_on_exit::mask);
+        exit_ctl_allow1(intel_x64::msrs::ia32_vmx_true_exit_ctls::acknowledge_interrupt_on_exit::mask);
         vm_exit_controls::acknowledge_interrupt_on_exit::enable();
         posted_interrupt_notification_vector::set(0x100UL);
     };
@@ -602,15 +542,15 @@ static void
 setup_check_control_vpid_checks_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::enable_vpid::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_vpid::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::enable_vpid::disable();
     };
@@ -618,8 +558,8 @@ setup_check_control_vpid_checks_paths(std::vector<struct control_flow_path> &cfg
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::enable_vpid::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_vpid::mask);
         secondary_processor_based_vm_execution_controls::enable_vpid::enable();
         virtual_processor_identifier::set(0U);
     };
@@ -655,8 +595,8 @@ setup_check_control_enable_ept_checks_paths(std::vector<struct control_flow_path
         proc_ctl2_allow1(secondary_processor_based_vm_execution_controls::enable_ept::mask);
         secondary_processor_based_vm_execution_controls::enable_ept::enable();
         ept_pointer::memory_type::set(vmcs::ept_pointer::memory_type::uncacheable);
-        g_msrs[ia32_vmx_ept_vpid_cap::addr] = ~(ia32_vmx_ept_vpid_cap::memory_type_uncacheable_supported::mask |
-                                                ia32_vmx_ept_vpid_cap::memory_type_write_back_supported::mask);
+        g_msrs[intel_x64::msrs::ia32_vmx_ept_vpid_cap::addr] = ~(intel_x64::msrs::ia32_vmx_ept_vpid_cap::memory_type_uncacheable_supported::mask |
+                                                intel_x64::msrs::ia32_vmx_ept_vpid_cap::memory_type_write_back_supported::mask);
     };
     path.throws_exception = true;
     cfg.push_back(path);
@@ -670,7 +610,7 @@ setup_check_control_enable_ept_checks_paths(std::vector<struct control_flow_path
     cfg.push_back(path);
 
     path.setup = [&] {
-        g_msrs[ia32_vmx_ept_vpid_cap::addr] = ia32_vmx_ept_vpid_cap::memory_type_write_back_supported::mask;
+        g_msrs[intel_x64::msrs::ia32_vmx_ept_vpid_cap::addr] = intel_x64::msrs::ia32_vmx_ept_vpid_cap::memory_type_write_back_supported::mask;
         ept_pointer::memory_type::set(vmcs::ept_pointer::memory_type::write_back);
         ept_pointer::page_walk_length_minus_one::set(0U);
     };
@@ -696,15 +636,15 @@ static void
 setup_check_control_enable_pml_checks_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::enable_pml::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_pml::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::enable_pml::disable();
     };
@@ -712,9 +652,9 @@ setup_check_control_enable_pml_checks_paths(std::vector<struct control_flow_path
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::enable_pml::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::enable_ept::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_pml::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_ept::mask);
         secondary_processor_based_vm_execution_controls::enable_pml::enable();
         secondary_processor_based_vm_execution_controls::enable_ept::disable();
     };
@@ -722,8 +662,8 @@ setup_check_control_enable_pml_checks_paths(std::vector<struct control_flow_path
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::enable_ept::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_ept::mask);
         secondary_processor_based_vm_execution_controls::enable_ept::enable();
         pml_address::set(0xff00000000000000U);
     };
@@ -743,15 +683,15 @@ static void
 setup_check_control_unrestricted_guests_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::unrestricted_guest::disable();
     };
@@ -759,9 +699,9 @@ setup_check_control_unrestricted_guests_paths(std::vector<struct control_flow_pa
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::enable_ept::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_ept::mask);
         secondary_processor_based_vm_execution_controls::unrestricted_guest::enable();
         secondary_processor_based_vm_execution_controls::enable_ept::disable();
     };
@@ -769,8 +709,8 @@ setup_check_control_unrestricted_guests_paths(std::vector<struct control_flow_pa
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::enable_ept::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_ept::mask);
         secondary_processor_based_vm_execution_controls::enable_ept::enable();
     };
     path.throws_exception = false;
@@ -783,15 +723,15 @@ setup_check_control_enable_vm_functions_paths(std::vector<struct control_flow_pa
     using namespace vmcs::secondary_processor_based_vm_execution_controls;
 
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::enable_vm_functions::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_vm_functions::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         enable_vm_functions::disable();
     };
@@ -799,7 +739,7 @@ setup_check_control_enable_vm_functions_paths(std::vector<struct control_flow_pa
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         proc_ctl2_disallow1(enable_vm_functions::mask);
         g_vmcs_fields[addr] |= enable_vm_functions::mask;
     };
@@ -807,10 +747,10 @@ setup_check_control_enable_vm_functions_paths(std::vector<struct control_flow_pa
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         proc_ctl2_allow1(enable_vm_functions::mask);
         vm_function_controls::set(1U);
-        g_msrs[ia32_vmx_vmfunc::addr] = 0;
+        g_msrs[intel_x64::msrs::ia32_vmx_vmfunc::addr] = 0;
     };
     path.throws_exception = true;
     cfg.push_back(path);
@@ -820,11 +760,11 @@ setup_check_control_enable_vm_functions_paths(std::vector<struct control_flow_pa
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::enable_vm_functions::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::enable_ept::mask);
-        vmfunc_ctl_allow1(ia32_vmx_vmfunc::eptp_switching::mask);
-        g_msrs[msrs::ia32_vmx_vmfunc::addr] |= (ia32_vmx_vmfunc::eptp_switching::mask << 32);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_vm_functions::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_ept::mask);
+        vmfunc_ctl_allow1(intel_x64::msrs::ia32_vmx_vmfunc::eptp_switching::mask);
+        g_msrs[intel_x64::msrs::ia32_vmx_vmfunc::addr] |= (intel_x64::msrs::ia32_vmx_vmfunc::eptp_switching::mask << 32);
         vm_function_controls::eptp_switching::enable();
         enable_ept::disable();
     };
@@ -832,7 +772,7 @@ setup_check_control_enable_vm_functions_paths(std::vector<struct control_flow_pa
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::enable_ept::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::enable_ept::mask);
         enable_ept::enable();
         secondary_processor_based_vm_execution_controls::enable_ept::enable();
         eptp_list_address::set_if_exists(0x0000000000000FFFULL);
@@ -857,15 +797,15 @@ static void
 setup_check_control_enable_vmcs_shadowing_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::vmcs_shadowing::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::vmcs_shadowing::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::vmcs_shadowing::disable();
     };
@@ -873,7 +813,7 @@ setup_check_control_enable_vmcs_shadowing_paths(std::vector<struct control_flow_
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::vmcs_shadowing::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::vmcs_shadowing::mask);
         secondary_processor_based_vm_execution_controls::vmcs_shadowing::enable();
         vmread_bitmap_address::set(1U);
     };
@@ -904,15 +844,15 @@ static void
 setup_check_control_enable_ept_violation_checks_paths(std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        proc_ctl_allow0(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl_allow0(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::disable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::ept_violation_ve::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::ept_violation_ve::mask);
         primary_processor_based_vm_execution_controls::activate_secondary_controls::enable();
         secondary_processor_based_vm_execution_controls::ept_violation_ve::disable();
     };
@@ -920,8 +860,8 @@ setup_check_control_enable_ept_violation_checks_paths(std::vector<struct control
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::virtual_interrupt_delivery::mask);
         secondary_processor_based_vm_execution_controls::ept_violation_ve::enable();
         virtualization_exception_information_address::set(1U);
     };
@@ -942,17 +882,17 @@ setup_check_control_vm_exit_ctls_reserved_properly_set_paths(std::vector<struct 
         &cfg)
 {
     path.setup = [&] {
-        g_msrs[ia32_vmx_true_exit_ctls::addr] = 0;
+        g_msrs[intel_x64::msrs::ia32_vmx_true_exit_ctls::addr] = 0;
         vm_exit_controls::set(0UL);
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
-    path.setup = [&] { g_msrs[ia32_vmx_true_exit_ctls::addr] = 1; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_exit_ctls::addr] = 1; };
     path.throws_exception = true;
     cfg.push_back(path);
 
-    path.setup = [&] { g_msrs[ia32_vmx_true_exit_ctls::addr] = 0xffffffff00000000UL; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_exit_ctls::addr] = 0xffffffff00000000UL; };
     path.throws_exception = false;
     cfg.push_back(path);
 }
@@ -962,15 +902,15 @@ setup_check_control_activate_and_save_preemption_timer_must_be_0_paths(
     std::vector<struct control_flow_path> &cfg)
 {
     path.setup = [&] {
-        pin_ctl_allow1(ia32_vmx_true_pinbased_ctls::activate_vmx_preemption_timer::mask);
+        pin_ctl_allow1(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::activate_vmx_preemption_timer::mask);
         pin_based_vm_execution_controls::activate_vmx_preemption_timer::enable();
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
     path.setup = [&] {
-        pin_ctl_allow0(ia32_vmx_true_pinbased_ctls::activate_vmx_preemption_timer::mask);
-        exit_ctl_allow1(ia32_vmx_true_exit_ctls::save_vmx_preemption_timer_value::mask);
+        pin_ctl_allow0(intel_x64::msrs::ia32_vmx_true_pinbased_ctls::activate_vmx_preemption_timer::mask);
+        exit_ctl_allow1(intel_x64::msrs::ia32_vmx_true_exit_ctls::save_vmx_preemption_timer_value::mask);
         pin_based_vm_execution_controls::activate_vmx_preemption_timer::disable();
         vm_exit_controls::save_vmx_preemption_timer_value::enable();
     };
@@ -978,7 +918,7 @@ setup_check_control_activate_and_save_preemption_timer_must_be_0_paths(
     cfg.push_back(path);
 
     path.setup = [&] {
-        exit_ctl_allow0(ia32_vmx_true_exit_ctls::save_vmx_preemption_timer_value::mask);
+        exit_ctl_allow0(intel_x64::msrs::ia32_vmx_true_exit_ctls::save_vmx_preemption_timer_value::mask);
         vm_exit_controls::save_vmx_preemption_timer_value::disable();
     };
     path.throws_exception = false;
@@ -1062,17 +1002,17 @@ setup_check_control_vm_entry_ctls_reserved_properly_set_paths(std::vector<struct
         &cfg)
 {
     path.setup = [&] {
-        g_msrs[ia32_vmx_true_entry_ctls::addr] = 0;
+        g_msrs[intel_x64::msrs::ia32_vmx_true_entry_ctls::addr] = 0;
         vm_entry_controls::set(0UL);
     };
     path.throws_exception = false;
     cfg.push_back(path);
 
-    path.setup = [&] { g_msrs[ia32_vmx_true_entry_ctls::addr] = 1; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_entry_ctls::addr] = 1; };
     path.throws_exception = true;
     cfg.push_back(path);
 
-    path.setup = [&] { g_msrs[ia32_vmx_true_entry_ctls::addr] = 0xffffffff00000000; };
+    path.setup = [&] { g_msrs[intel_x64::msrs::ia32_vmx_true_entry_ctls::addr] = 0xffffffff00000000; };
     path.throws_exception = false;
     cfg.push_back(path);
 }
@@ -1096,7 +1036,7 @@ setup_check_control_event_injection_type_vector_checks_paths(std::vector<struct 
 
     path.setup = [&] {
         interruption_type::set(interruption_type::other_event);
-        g_msrs[ia32_vmx_true_procbased_ctls::addr] = 0;
+        g_msrs[intel_x64::msrs::ia32_vmx_true_procbased_ctls::addr] = 0;
     };
     path.throws_exception = true;
     cfg.push_back(path);
@@ -1114,7 +1054,7 @@ setup_check_control_event_injection_type_vector_checks_paths(std::vector<struct 
 
     path.setup = [&] {
         interruption_type::set(interruption_type::other_event);
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::monitor_trap_flag::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::monitor_trap_flag::mask);
     };
     path.throws_exception = true;
     cfg.push_back(path);
@@ -1135,8 +1075,8 @@ setup_check_control_event_injection_delivery_ec_checks_paths(std::vector<struct 
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow1(ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow1(intel_x64::msrs::ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
         valid_bit::enable();
         deliver_error_code_bit::enable();
         guest_cr0::set(0U);
@@ -1146,8 +1086,8 @@ setup_check_control_event_injection_delivery_ec_checks_paths(std::vector<struct 
     cfg.push_back(path);
 
     path.setup = [&] {
-        proc_ctl_allow1(ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
-        proc_ctl2_allow0(ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
+        proc_ctl_allow1(intel_x64::msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::mask);
+        proc_ctl2_allow0(intel_x64::msrs::ia32_vmx_procbased_ctls2::unrestricted_guest::mask);
         secondary_processor_based_vm_execution_controls::unrestricted_guest::disable();
         interruption_type::set(interruption_type::non_maskable_interrupt);
     };
@@ -1242,7 +1182,7 @@ setup_check_control_event_injection_instr_length_checks_paths(std::vector<struct
     path.setup = [&] {
         interruption_type::set(interruption_type::software_interrupt);
         vm_entry_instruction_length::set(0UL);
-        g_msrs[ia32_vmx_misc::addr] = 0;
+        g_msrs[intel_x64::msrs::ia32_vmx_misc::addr] = 0;
     };
     path.throws_exception = true;
     cfg.push_back(path);
@@ -1575,24 +1515,22 @@ TEST_CASE("check_control_entry_msr_load_address")
 TEST_CASE("check_control_reserved_properly_set")
 {
     MockRepository mocks;
-    auto mm = mocks.Mock<memory_manager_x64>();
-    setup_intrinsics(mocks, mm);
+    setup_mm(mocks);
 
-    g_msrs[ia32_vmx_true_entry_ctls::addr] = 0xffffffff00000000UL;
+    g_msrs[intel_x64::msrs::ia32_vmx_true_entry_ctls::addr] = 0xffffffff00000000UL;
     vm_entry_controls::set(0x1234UL);
 
-    auto msr_addr = ia32_vmx_true_entry_ctls::addr;
+    auto msr_addr = intel_x64::msrs::ia32_vmx_true_entry_ctls::addr;
     auto ctls = vm_entry_controls::get();
     auto name = vm_entry_controls::name;
 
-    CHECK_NOTHROW(control_reserved_properly_set(msr_addr, ctls, name));
+    CHECK_NOTHROW(check::control_reserved_properly_set(msr_addr, ctls, name));
 }
 
 TEST_CASE("check_memory_type_reserved")
 {
     MockRepository mocks;
-    auto mm = mocks.Mock<memory_manager_x64>();
-    setup_intrinsics(mocks, mm);
+    setup_mm(mocks);
 
     CHECK_FALSE(memory_type_reserved(x64::memory_type::write_through));
     CHECK(memory_type_reserved(8U));
