--- bfvmm/tests/hve/arch/intel_x64/vmcs/test_vmcs_promote_mock.cpp
@@ -1,25 +1,0 @@
-//
-// Bareflank Hypervisor
-// Copyright (C) 2015 Assured Information Security, Inc.
-//
-// This library is free software; you can redistribute it and/or
-// modify it under the terms of the GNU Lesser General Public
-// License as published by the Free Software Foundation; either
-// version 2.1 of the License, or (at your option) any later version.
-//
-// This library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-// Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public
-// License along with this library; if not, write to the Free Software
-// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
-
-#include <catch/catch.hpp>
-#include <hve/arch/intel_x64/vmcs/vmcs_intel_x64_promote.h>
-
-TEST_CASE("")
-{
-    CHECK_NOTHROW(vmcs_promote(0, 0));
-}
