+++ bfvmm/tests/hve/arch/intel_x64/vmcs/test_vmcs_state_hvm.cpp
@@ -16,240 +16,15 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include <catch/catch.hpp>
-#include <hippomocks.h>
-
-#include <intrinsics/x86/intel_x64.h>
-#include <vmcs/vmcs_intel_x64_host_vm_state.h>
+#include <support/arch/intel_x64/test_support.h>
 
 #ifdef _HIPPOMOCKS__ENABLE_CFUNC_MOCKING_SUPPORT
-
-using namespace x64;
-
-uint16_t test_es;
-uint16_t test_cs;
-uint16_t test_ss;
-uint16_t test_ds;
-uint16_t test_fs;
-uint16_t test_gs;
-uint16_t test_ldtr;
-uint16_t test_tr;
-
-uint64_t test_cr0;
-uint64_t test_cr3;
-uint64_t test_cr4;
-uint64_t test_dr7;
-uint64_t test_rflags;
-
-gdt_reg_x64_t test_gdtr{};
-idt_reg_x64_t test_idtr{};
-
-std::vector<gdt_x64::segment_descriptor_type> test_gdt = {
-    0x0,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF
-};
-
-std::vector<idt_x64::interrupt_descriptor_type> test_idt{512};
-
-std::map<uint32_t, uint64_t> g_msrs;
-std::map<uint32_t, uint32_t> g_eax_cpuid;
-
-static uint16_t
-test_read_es() noexcept
-{ return test_es; }
-
-static uint16_t
-test_read_cs() noexcept
-{ return test_cs; }
-
-static uint16_t
-test_read_ss() noexcept
-{ return test_ss; }
-
-static uint16_t
-test_read_ds() noexcept
-{ return test_ds; }
-
-static uint16_t
-test_read_fs() noexcept
-{ return test_fs; }
-
-static uint16_t
-test_read_gs() noexcept
-{ return test_gs; }
-
-static uint16_t
-test_read_tr() noexcept
-{ return test_tr; }
-
-static uint16_t
-test_read_ldtr() noexcept
-{ return test_ldtr; }
-
-static void
-test_write_es(uint16_t val) noexcept
-{ test_es = val; }
-
-static void
-test_write_cs(uint16_t val) noexcept
-{ test_cs = val; }
-
-static void
-test_write_ss(uint16_t val) noexcept
-{ test_ss = val; }
-
-static void
-test_write_ds(uint16_t val) noexcept
-{ test_ds = val; }
-
-static void
-test_write_fs(uint16_t val) noexcept
-{ test_fs = val; }
-
-static void
-test_write_gs(uint16_t val) noexcept
-{ test_gs = val; }
-
-static void
-test_write_tr(uint16_t val) noexcept
-{ test_tr = val; }
-
-static void
-test_write_ldtr(uint16_t val) noexcept
-{ test_ldtr = val; }
-
-static uint64_t
-test_read_cr0() noexcept
-{ return test_cr0; }
-
-static uint64_t
-test_read_cr3() noexcept
-{ return test_cr3; }
-
-static uint64_t
-test_read_cr4() noexcept
-{ return test_cr4; }
-
-static void
-test_write_cr0(uint64_t val) noexcept
-{ test_cr0 = val; }
-
-static void
-test_write_cr3(uint64_t val) noexcept
-{ test_cr3 = val; }
-
-static void
-test_write_cr4(uint64_t val) noexcept
-{ test_cr4 = val; }
-
-static uint64_t
-test_read_dr7() noexcept
-{ return test_dr7; }
-
-static void
-test_write_dr7(uint64_t val) noexcept
-{ test_dr7 = val; }
-
-static uint64_t
-test_read_rflags() noexcept
-{ return test_rflags; }
-
-static void
-test_write_rflags(uint64_t val) noexcept
-{ test_rflags = val; }
-
-static void
-test_read_gdt(gdt_reg_x64_t *gdt_reg) noexcept
-{ *gdt_reg = test_gdtr; }
-
-static void
-test_read_idt(idt_reg_x64_t *idt_reg) noexcept
-{ *idt_reg = test_idtr; }
-
-static uint64_t
-test_read_msr(uint32_t addr) noexcept
-{ return g_msrs[addr]; }
-
-static void
-test_write_msr(uint32_t addr, uint64_t val) noexcept
-{ g_msrs[addr] = val; }
-
-static uint32_t
-test_cpuid_eax(uint32_t val) noexcept
-{ return g_eax_cpuid[val]; }
-
-void
-setup_gdt()
-{
-    auto limit = test_gdt.size() * sizeof(gdt_x64::segment_descriptor_type);
-
-    test_gdtr.base = &test_gdt.at(0);
-    test_gdtr.limit = gsl::narrow_cast<gdt_reg_x64_t::limit_type>(limit);
-}
-
-void
-setup_idt()
-{
-    auto limit = test_idt.size() * sizeof(idt_x64::interrupt_descriptor_type);
-
-    test_idtr.base = &test_idt.at(0);
-    test_idtr.limit = gsl::narrow_cast<idt_reg_x64_t::limit_type>(limit);
-}
-
-static void
-setup_intrinsics(MockRepository &mocks)
-{
-    mocks.OnCallFunc(_read_es).Do(test_read_es);
-    mocks.OnCallFunc(_read_cs).Do(test_read_cs);
-    mocks.OnCallFunc(_read_ss).Do(test_read_ss);
-    mocks.OnCallFunc(_read_ds).Do(test_read_ds);
-    mocks.OnCallFunc(_read_fs).Do(test_read_fs);
-    mocks.OnCallFunc(_read_gs).Do(test_read_gs);
-    mocks.OnCallFunc(_read_tr).Do(test_read_tr);
-    mocks.OnCallFunc(_read_ldtr).Do(test_read_ldtr);
-
-    mocks.OnCallFunc(_write_es).Do(test_write_es);
-    mocks.OnCallFunc(_write_cs).Do(test_write_cs);
-    mocks.OnCallFunc(_write_ss).Do(test_write_ss);
-    mocks.OnCallFunc(_write_ds).Do(test_write_ds);
-    mocks.OnCallFunc(_write_fs).Do(test_write_fs);
-    mocks.OnCallFunc(_write_gs).Do(test_write_gs);
-    mocks.OnCallFunc(_write_tr).Do(test_write_tr);
-    mocks.OnCallFunc(_write_ldtr).Do(test_write_ldtr);
-
-    mocks.OnCallFunc(_read_cr0).Do(test_read_cr0);
-    mocks.OnCallFunc(_read_cr3).Do(test_read_cr3);
-    mocks.OnCallFunc(_read_cr4).Do(test_read_cr4);
-    mocks.OnCallFunc(_write_cr0).Do(test_write_cr0);
-    mocks.OnCallFunc(_write_cr3).Do(test_write_cr3);
-    mocks.OnCallFunc(_write_cr4).Do(test_write_cr4);
-
-    mocks.OnCallFunc(_read_dr7).Do(test_read_dr7);
-    mocks.OnCallFunc(_write_dr7).Do(test_write_dr7);
-
-    mocks.OnCallFunc(_read_rflags).Do(test_read_rflags);
-    mocks.OnCallFunc(_write_rflags).Do(test_write_rflags);
-
-    mocks.OnCallFunc(_read_gdt).Do(test_read_gdt);
-    mocks.OnCallFunc(_read_idt).Do(test_read_idt);
-
-    mocks.OnCallFunc(_read_msr).Do(test_read_msr);
-    mocks.OnCallFunc(_write_msr).Do(test_write_msr);
-
-    mocks.OnCallFunc(_cpuid_eax).Do(test_cpuid_eax);
-
-    setup_gdt();
-    setup_idt();
-}
 
 TEST_CASE("vmcs: host_vm_state")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     CHECK_NOTHROW(vmcs_intel_x64_host_vm_state{});
 }
@@ -257,7 +32,8 @@ TEST_CASE("vmcs: host_vm_state")
 TEST_CASE("vmcs: host_vm_state_segment_registers")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     segment_register::es::set(1U);
     segment_register::cs::set(1U);
@@ -283,7 +59,8 @@ TEST_CASE("vmcs: host_vm_state_segment_registers")
 TEST_CASE("vmcs: host_vm_state_control_registers")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     intel_x64::cr0::set(1U);
     intel_x64::cr3::set(2U);
@@ -299,7 +76,8 @@ TEST_CASE("vmcs: host_vm_state_control_registers")
 TEST_CASE("vmcs: host_vm_state_debug_registers")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     dr7::set(42U);
     vmcs_intel_x64_host_vm_state state{};
@@ -310,7 +88,8 @@ TEST_CASE("vmcs: host_vm_state_debug_registers")
 TEST_CASE("vmcs: host_vm_state_rflags")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     rflags::set(42U);
     vmcs_intel_x64_host_vm_state state{};
@@ -321,44 +100,49 @@ TEST_CASE("vmcs: host_vm_state_rflags")
 TEST_CASE("vmcs: host_vm_state_gdt_base")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     vmcs_intel_x64_host_vm_state state{};
-    CHECK(state.gdt_base() == bfrcast(gdt_x64::integer_pointer, test_gdtr.base));
+    CHECK(state.gdt_base() == bfscast(gdt_x64::integer_pointer, g_gdtr.base));
 }
 
 TEST_CASE("vmcs: host_vm_state_idt_base")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     vmcs_intel_x64_host_vm_state state{};
-    CHECK(state.idt_base() == bfrcast(gdt_x64::integer_pointer, test_idtr.base));
+    CHECK(state.idt_base() == bfscast(gdt_x64::integer_pointer, g_idtr.base));
 }
 
 TEST_CASE("vmcs: host_vm_state_gdt_limit")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     vmcs_intel_x64_host_vm_state state{};
-    CHECK(state.gdt_limit() == test_gdtr.limit);
+    CHECK(state.gdt_limit() == g_gdtr.limit);
 }
 
 TEST_CASE("vmcs: host_vm_state_idt_limit")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     vmcs_intel_x64_host_vm_state state{};
-    CHECK(state.idt_limit() == test_idtr.limit);
+    CHECK(state.idt_limit() == g_idtr.limit);
 }
 
 TEST_CASE("vmcs: host_vm_state_es_limit")
 {
     SECTION("es_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::es::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -368,7 +152,8 @@ TEST_CASE("vmcs: host_vm_state_es_limit")
 
     SECTION("es_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::es::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -381,7 +166,8 @@ TEST_CASE("vmcs: host_vm_state_cs_limit")
 {
     SECTION("cs_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::cs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -391,7 +177,8 @@ TEST_CASE("vmcs: host_vm_state_cs_limit")
 
     SECTION("cs_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::cs::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -404,7 +191,8 @@ TEST_CASE("vmcs: host_vm_state_ss_limit")
 {
     SECTION("ss_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ss::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -414,7 +202,8 @@ TEST_CASE("vmcs: host_vm_state_ss_limit")
 
     SECTION("ss_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ss::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -427,7 +216,8 @@ TEST_CASE("vmcs: host_vm_state_ds_limit")
 {
     SECTION("ds_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ds::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -437,7 +227,8 @@ TEST_CASE("vmcs: host_vm_state_ds_limit")
 
     SECTION("ds_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ds::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -450,7 +241,8 @@ TEST_CASE("vmcs: host_vm_state_fs_limit")
 {
     SECTION("fs_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::fs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -460,7 +252,8 @@ TEST_CASE("vmcs: host_vm_state_fs_limit")
 
     SECTION("fs_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::fs::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -473,7 +266,8 @@ TEST_CASE("vmcs: host_vm_state_gs_limit")
 {
     SECTION("gs_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::gs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -483,7 +277,8 @@ TEST_CASE("vmcs: host_vm_state_gs_limit")
 
     SECTION("gs_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::gs::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -496,7 +291,8 @@ TEST_CASE("vmcs: host_vm_state_tr_limit")
 {
     SECTION("tr_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::tr::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -506,7 +302,8 @@ TEST_CASE("vmcs: host_vm_state_tr_limit")
 
     SECTION("tr_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::tr::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -519,7 +316,8 @@ TEST_CASE("vmcs: host_vm_state_ldtr_limit")
 {
     SECTION("ldtr_limit == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ldtr::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -529,7 +327,8 @@ TEST_CASE("vmcs: host_vm_state_ldtr_limit")
 
     SECTION("ldtr_limit == 0xFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ldtr::index::set(1U);
         vmcs_intel_x64_host_vm_state state{};
@@ -542,7 +341,8 @@ TEST_CASE("vmcs: host_vm_state_es_access_rights")
 {
     SECTION("es_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::es::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -552,7 +352,8 @@ TEST_CASE("vmcs: host_vm_state_es_access_rights")
 
     SECTION("es_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::es::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -565,7 +366,8 @@ TEST_CASE("vmcs: host_vm_state_cs_access_rights")
 {
     SECTION("cs_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::cs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -575,7 +377,8 @@ TEST_CASE("vmcs: host_vm_state_cs_access_rights")
 
     SECTION("cs_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::cs::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -588,7 +391,8 @@ TEST_CASE("vmcs: host_vm_state_ss_access_rights")
 {
     SECTION("ss_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ss::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -598,7 +402,8 @@ TEST_CASE("vmcs: host_vm_state_ss_access_rights")
 
     SECTION("ss_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ss::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -611,7 +416,8 @@ TEST_CASE("vmcs: host_vm_state_ds_access_rights")
 {
     SECTION("ds_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ds::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -621,7 +427,8 @@ TEST_CASE("vmcs: host_vm_state_ds_access_rights")
 
     SECTION("ds_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ds::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -634,7 +441,8 @@ TEST_CASE("vmcs: host_vm_state_fs_access_rights")
 {
     SECTION("fs_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::fs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -644,7 +452,8 @@ TEST_CASE("vmcs: host_vm_state_fs_access_rights")
 
     SECTION("fs_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::fs::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -657,7 +466,8 @@ TEST_CASE("vmcs: host_vm_state_gs_access_rights")
 {
     SECTION("gs_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::gs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -667,7 +477,8 @@ TEST_CASE("vmcs: host_vm_state_gs_access_rights")
 
     SECTION("gs_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::gs::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -680,7 +491,8 @@ TEST_CASE("vmcs: host_vm_state_tr_access_rights")
 {
     SECTION("tr_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::tr::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -690,7 +502,8 @@ TEST_CASE("vmcs: host_vm_state_tr_access_rights")
 
     SECTION("tr_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::tr::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -703,7 +516,8 @@ TEST_CASE("vmcs: host_vm_state_ldtr_access_rights")
 {
     SECTION("ldtr_access_rights unusable") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ldtr::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -713,7 +527,8 @@ TEST_CASE("vmcs: host_vm_state_ldtr_access_rights")
 
     SECTION("ldtr_access_rights == 0x70") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ldtr::index::set(2U);
         vmcs_intel_x64_host_vm_state state{};
@@ -726,7 +541,8 @@ TEST_CASE("vmcs: host_vm_state_es_base")
 {
     SECTION("es_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::es::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -736,7 +552,8 @@ TEST_CASE("vmcs: host_vm_state_es_base")
 
     SECTION("es_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::es::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -749,7 +566,8 @@ TEST_CASE("vmcs: host_vm_state_cs_base")
 {
     SECTION("cs_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::cs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -759,7 +577,8 @@ TEST_CASE("vmcs: host_vm_state_cs_base")
 
     SECTION("cs_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::cs::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -772,7 +591,8 @@ TEST_CASE("vmcs: host_vm_state_ss_base")
 {
     SECTION("ss_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ss::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -782,7 +602,8 @@ TEST_CASE("vmcs: host_vm_state_ss_base")
 
     SECTION("ss_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ss::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -795,7 +616,8 @@ TEST_CASE("vmcs: host_vm_state_ds_base")
 {
     SECTION("ds_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ds::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -805,7 +627,8 @@ TEST_CASE("vmcs: host_vm_state_ds_base")
 
     SECTION("ds_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ds::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -818,7 +641,8 @@ TEST_CASE("vmcs: host_vm_state_fs_base")
 {
     SECTION("fs_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::fs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -828,7 +652,8 @@ TEST_CASE("vmcs: host_vm_state_fs_base")
 
     SECTION("fs_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::fs::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -841,7 +666,8 @@ TEST_CASE("vmcs: host_vm_state_gs_base")
 {
     SECTION("gs_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::gs::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -851,7 +677,8 @@ TEST_CASE("vmcs: host_vm_state_gs_base")
 
     SECTION("gs_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::gs::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -864,7 +691,8 @@ TEST_CASE("vmcs: host_vm_state_tr_base")
 {
     SECTION("tr_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::tr::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -874,7 +702,8 @@ TEST_CASE("vmcs: host_vm_state_tr_base")
 
     SECTION("tr_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::tr::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -887,7 +716,8 @@ TEST_CASE("vmcs: host_vm_state_ldtr_base")
 {
     SECTION("ldtr_base == 0") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ldtr::index::set(0U);
         vmcs_intel_x64_host_vm_state state{};
@@ -897,7 +727,8 @@ TEST_CASE("vmcs: host_vm_state_ldtr_base")
 
     SECTION("ldtr_base == 0xFFFFFFFF") {
         MockRepository mocks;
-        setup_intrinsics(mocks);
+        setup_gdt();
+    setup_idt();
 
         segment_register::ldtr::index::set(3U);
         vmcs_intel_x64_host_vm_state state{};
@@ -909,7 +740,8 @@ TEST_CASE("vmcs: host_vm_state_ldtr_base")
 TEST_CASE("vmcs: host_vm_state_ia32_msrs_no_perf")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     intel_x64::msrs::ia32_debugctl::set(42U);
     x64::msrs::ia32_pat::set(42U);
@@ -937,7 +769,8 @@ TEST_CASE("vmcs: host_vm_state_ia32_msrs_no_perf")
 TEST_CASE("vmcs: host_vm_state_ia32_msrs_perf")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     intel_x64::msrs::ia32_debugctl::set(42U);
     x64::msrs::ia32_pat::set(42U);
@@ -966,7 +799,8 @@ TEST_CASE("vmcs: host_vm_state_ia32_msrs_perf")
 TEST_CASE("vmcs: host_vm_state_dump")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     vmcs_intel_x64_host_vm_state state{};
     CHECK_NOTHROW(state.dump());
