+++ bfvmm/tests/hve/arch/intel_x64/vmcs/test_vmcs_state_vmm.cpp
@@ -16,55 +16,17 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include <catch/catch.hpp>
-#include <hippomocks.h>
-
-#include <intrinsics/x86/intel_x64.h>
-#include <vmcs/vmcs_intel_x64_vmm_state.h>
-
-#include <memory_manager/pat_x64.h>
-#include <memory_manager/root_page_table_x64.h>
+#include <support/arch/intel_x64/test_support.h>
 
 #ifdef _HIPPOMOCKS__ENABLE_CFUNC_MOCKING_SUPPORT
-
-using namespace x64;
-
-static uint32_t
-test_cpuid_ecx(uint32_t addr) noexcept
-{
-    bfignored(addr);
-    return 0xFFFFFFFFU;
-}
-
-static uint32_t
-test_cpuid_subebx(uint32_t addr, uint32_t leaf)
-{
-    bfignored(addr);
-    bfignored(leaf);
-    return 0xFFFFFFFFU;
-}
-
-static void
-setup_mm(MockRepository &mocks)
-{
-    auto pt = mocks.Mock<root_page_table_x64>();
-
-    mocks.OnCallFunc(root_pt).Return(pt);
-    mocks.OnCall(pt, root_page_table_x64::cr3).Return(42U);
-}
-
-static void
-setup_intrinsics(MockRepository &mocks)
-{
-    mocks.OnCallFunc(_cpuid_ecx).Do(test_cpuid_ecx);
-    mocks.OnCallFunc(_cpuid_subebx).Do(test_cpuid_subebx);
-}
 
 TEST_CASE("vmcs: vmm_state")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
+
+    g_ecx_cpuid[intel_x64::cpuid::feature_information::addr] = 0xFFFFFFFF;
+    g_ebx_cpuid[intel_x64::cpuid::extended_feature_flags::addr] = 0xFFFFFFFF;
 
     vmcs_intel_x64_vmm_state state{};
 }
@@ -73,7 +35,6 @@ TEST_CASE("vmcs: state_segment_registers")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
 
@@ -88,7 +49,6 @@ TEST_CASE("vmcs: state_control_registers")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
 
@@ -101,7 +61,6 @@ TEST_CASE("vmcs: state_debug_registers")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
     CHECK(state.dr7() == 0U);
@@ -111,7 +70,6 @@ TEST_CASE("vmcs: state_rflags")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
     CHECK(state.rflags() == 0U);
@@ -121,7 +79,6 @@ TEST_CASE("vmcs: state_gdt_base")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
     CHECK(state.gdt_base() != 0U);
@@ -131,7 +88,6 @@ TEST_CASE("vmcs: state_idt_base")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
     CHECK(state.idt_base() != 0U);
@@ -141,7 +97,6 @@ TEST_CASE("vmcs: state_gdt_limit")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
 
@@ -152,7 +107,6 @@ TEST_CASE("vmcs: state_idt_limit")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
     CHECK(state.idt_limit() == 2047U);
@@ -162,7 +116,6 @@ TEST_CASE("vmcs: state_segment_registers_limit")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
 
@@ -177,7 +130,6 @@ TEST_CASE("vmcs: state_segment_registers_access_rights")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
 
@@ -192,7 +144,6 @@ TEST_CASE("vmcs: state_segment_register_base")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
 
@@ -207,7 +158,6 @@ TEST_CASE("vmcs: state_msrs")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
 
@@ -219,7 +169,6 @@ TEST_CASE("vmcs: state_dump")
 {
     MockRepository mocks;
     setup_mm(mocks);
-    setup_intrinsics(mocks);
 
     vmcs_intel_x64_vmm_state state{};
     CHECK_NOTHROW(state.dump());
