+++ bfvmm/tests/vcpu/arch/intel_x64/test_vcpu.cpp
@@ -16,194 +16,9 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include <catch/catch.hpp>
-#include <hippomocks.h>
-
-#include <cstdlib>
-
-#include <vcpu/vcpu_intel_x64.h>
-#include <memory_manager/memory_manager_x64.h>
-#include <memory_manager/root_page_table_x64.h>
-
-#include <intrinsics/x86/common_x64.h>
-using namespace x64;
+#include <support/arch/intel_x64/test_support.h>
 
 #ifdef _HIPPOMOCKS__ENABLE_CFUNC_MOCKING_SUPPORT
-
-gdt_reg_x64_t test_gdtr{};
-idt_reg_x64_t test_idtr{};
-
-std::vector<gdt_x64::segment_descriptor_type> test_gdt = {
-    0x0,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF,
-    0xFF7FFFFFFFFFFFFF
-};
-
-std::vector<idt_x64::interrupt_descriptor_type> test_idt{512};
-
-void
-setup_gdt()
-{
-    auto limit = test_gdt.size() * sizeof(gdt_x64::segment_descriptor_type);
-
-    test_gdtr.base = &test_gdt.at(0);
-    test_gdtr.limit = gsl::narrow_cast<gdt_reg_x64_t::limit_type>(limit);
-}
-
-void
-setup_idt()
-{
-    auto limit = test_idt.size() * sizeof(idt_x64::interrupt_descriptor_type);
-
-    test_idtr.base = &test_idt.at(0);
-    test_idtr.limit = gsl::narrow_cast<idt_reg_x64_t::limit_type>(limit);
-}
-
-static uint64_t
-test_read_msr(uint32_t addr) noexcept
-{ bfignored(addr); return 0; }
-
-static uint64_t
-test_read_cr0() noexcept
-{ return 0; }
-
-static uint64_t
-test_read_cr3() noexcept
-{ return 0; }
-
-static uint64_t
-test_read_cr4() noexcept
-{ return 0; }
-
-static uint64_t
-test_read_rflags() noexcept
-{ return 0; }
-
-static uint64_t
-test_read_dr7() noexcept
-{ return 0; }
-
-static void
-test_read_gdt(gdt_reg_x64_t *gdt_reg) noexcept
-{ *gdt_reg = test_gdtr; }
-
-static void
-test_read_idt(idt_reg_x64_t *idt_reg) noexcept
-{ *idt_reg = test_idtr; }
-
-static uint16_t
-test_read_es() noexcept
-{ return 0; }
-
-static uint16_t
-test_read_cs() noexcept
-{ return 0; }
-
-static uint16_t
-test_read_ss() noexcept
-{ return 0; }
-
-static uint16_t
-test_read_ds() noexcept
-{ return 0; }
-
-static uint16_t
-test_read_fs() noexcept
-{ return 0; }
-
-static uint16_t
-test_read_gs() noexcept
-{ return 0; }
-
-static uint16_t
-test_read_ldtr() noexcept
-{ return 0; }
-
-static uint16_t
-test_read_tr() noexcept
-{ return 0; }
-
-static uint32_t
-test_cpuid_ecx(uint32_t val) noexcept
-{ bfignored(val); return 0x04000000U; }
-
-static uint32_t
-test_cpuid_eax(uint32_t val) noexcept
-{ bfignored(val); return 0x00000000U; }
-
-static uint32_t
-test_cpuid_subebx(uint32_t addr, uint32_t leaf)
-{
-    bfignored(addr);
-    bfignored(leaf);
-
-    return 0x00000000U;
-}
-
-static void
-setup_intrinsics(MockRepository &mocks)
-{
-    mocks.OnCallFunc(_read_msr).Do(test_read_msr);
-    mocks.OnCallFunc(_read_cr0).Do(test_read_cr0);
-    mocks.OnCallFunc(_read_cr3).Do(test_read_cr3);
-    mocks.OnCallFunc(_read_cr4).Do(test_read_cr4);
-    mocks.OnCallFunc(_read_rflags).Do(test_read_rflags);
-    mocks.OnCallFunc(_read_dr7).Do(test_read_dr7);
-    mocks.OnCallFunc(_read_gdt).Do(test_read_gdt);
-    mocks.OnCallFunc(_read_idt).Do(test_read_idt);
-    mocks.OnCallFunc(_read_es).Do(test_read_es);
-    mocks.OnCallFunc(_read_cs).Do(test_read_cs);
-    mocks.OnCallFunc(_read_ss).Do(test_read_ss);
-    mocks.OnCallFunc(_read_ds).Do(test_read_ds);
-    mocks.OnCallFunc(_read_fs).Do(test_read_fs);
-    mocks.OnCallFunc(_read_gs).Do(test_read_gs);
-    mocks.OnCallFunc(_read_ldtr).Do(test_read_ldtr);
-    mocks.OnCallFunc(_read_tr).Do(test_read_tr);
-    mocks.OnCallFunc(_cpuid_ecx).Do(test_cpuid_ecx);
-    mocks.OnCallFunc(_cpuid_eax).Do(test_cpuid_eax);
-    mocks.OnCallFunc(_cpuid_subebx).Do(test_cpuid_subebx);
-
-    setup_gdt();
-    setup_idt();
-}
-
-static auto
-setup_mm(MockRepository &mocks)
-{
-    auto mm = mocks.Mock<memory_manager_x64>();
-    mocks.OnCallFunc(memory_manager_x64::instance).Return(mm);
-    mocks.OnCall(mm, memory_manager_x64::virtptr_to_physint).Return(0x000000ABCDEF0000);
-
-    return mm;
-}
-
-static auto
-setup_pt(MockRepository &mocks)
-{
-    auto pt = mocks.Mock<root_page_table_x64>();
-    mocks.OnCallFunc(root_pt).Return(pt);
-    mocks.OnCall(pt, root_page_table_x64::cr3).Return(0x000000ABCDEF0000);
-
-    return pt;
-}
-
-template<typename T> auto
-mock_no_delete(MockRepository &mocks)
-{
-    auto ptr = mocks.Mock<T>();
-    mocks.OnCallDestructor(ptr);
-
-    return ptr;
-}
-
-template <typename T> auto
-mock_unique(MockRepository &mocks)
-{
-    return std::unique_ptr<T>(mock_no_delete<T>(mocks));
-}
 
 TEST_CASE("vcpu_intel_x64: invalid_id")
 {
@@ -213,7 +28,8 @@ TEST_CASE("vcpu_intel_x64: invalid_id")
 TEST_CASE("vcpu_intel_x64: valid")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
 
     auto test = [&] {
         auto on = mock_unique<vmxon_intel_x64>(mocks);
@@ -237,7 +53,8 @@ TEST_CASE("vcpu_intel_x64: valid")
 TEST_CASE("vcpu_intel_x64: init_null_params")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -256,7 +73,8 @@ TEST_CASE("vcpu_intel_x64: init_null_params")
 TEST_CASE("vcpu_intel_x64: init_valid_params")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -287,7 +105,8 @@ TEST_CASE("vcpu_intel_x64: init_valid_params")
 TEST_CASE("vcpu_intel_x64: init_valid")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -318,7 +137,8 @@ TEST_CASE("vcpu_intel_x64: init_valid")
 TEST_CASE("vcpu_intel_x64: init_vmcs_throws")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -349,7 +169,8 @@ TEST_CASE("vcpu_intel_x64: init_vmcs_throws")
 TEST_CASE("vcpu_intel_x64: fini_null_params")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -368,7 +189,8 @@ TEST_CASE("vcpu_intel_x64: fini_null_params")
 TEST_CASE("vcpu_intel_x64: fini_valid_params")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -400,7 +222,8 @@ TEST_CASE("vcpu_intel_x64: fini_valid_params")
 TEST_CASE("vcpu_intel_x64: fini_valid")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -432,7 +255,8 @@ TEST_CASE("vcpu_intel_x64: fini_valid")
 TEST_CASE("vcpu_intel_x64: fini_no_init")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -457,7 +281,8 @@ TEST_CASE("vcpu_intel_x64: fini_no_init")
 TEST_CASE("vcpu_intel_x64: run_launch")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -495,7 +320,8 @@ TEST_CASE("vcpu_intel_x64: run_launch")
 TEST_CASE("vcpu_intel_x64: run_launch_is_host_vcpu")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -533,7 +359,8 @@ TEST_CASE("vcpu_intel_x64: run_launch_is_host_vcpu")
 TEST_CASE("vcpu_intel_x64: run_resume")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -572,7 +399,8 @@ TEST_CASE("vcpu_intel_x64: run_resume")
 TEST_CASE("vcpu_intel_x64: run_no_init")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -609,7 +437,8 @@ TEST_CASE("vcpu_intel_x64: run_no_init")
 TEST_CASE("vcpu_intel_x64: run_vmxon_throws")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -647,7 +476,8 @@ TEST_CASE("vcpu_intel_x64: run_vmxon_throws")
 TEST_CASE("vcpu_intel_x64: run_vmcs_throws")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -685,7 +515,8 @@ TEST_CASE("vcpu_intel_x64: run_vmcs_throws")
 TEST_CASE("vcpu_intel_x64: hlt_no_init")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -722,7 +553,8 @@ TEST_CASE("vcpu_intel_x64: hlt_no_init")
 TEST_CASE("vcpu_intel_x64: hlt_no_run")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -760,7 +592,8 @@ TEST_CASE("vcpu_intel_x64: hlt_no_run")
 TEST_CASE("vcpu_intel_x64: hlt_valid")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -799,7 +632,8 @@ TEST_CASE("vcpu_intel_x64: hlt_valid")
 TEST_CASE("vcpu_intel_x64: hlt_valid_is_host_vcpu")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
 
@@ -838,7 +672,8 @@ TEST_CASE("vcpu_intel_x64: hlt_valid_is_host_vcpu")
 TEST_CASE("vcpu_intel_x64: hlt_vmxon_throws")
 {
     MockRepository mocks;
-    setup_intrinsics(mocks);
+    setup_gdt();
+    setup_idt();
     setup_mm(mocks);
     setup_pt(mocks);
