+++ scripts/cmake/flags/test_flags.cmake
@@ -28,6 +28,7 @@ list(APPEND BFFLAGS_TEST
     -D${OSTYPE}
     -D${ABITYPE}
     -DENABLE_BUILD_TEST
+    -DDEBUG_LEVEL=5
 )
 
 if(NOT WIN32)
