+++ scripts/cmake/macros.cmake
@@ -217,18 +217,6 @@ function(generate_flags PREFIX)
             list(APPEND _C_FLAGS ${BFFLAGS_USERSPACE_AARCH64})
             list(APPEND _CXX_FLAGS ${BFFLAGS_USERSPACE_AARCH64})
         endif()
-        if(ENABLE_ASAN)
-            list(APPEND _C_FLAGS ${BFFLAGS_ASAN})
-            list(APPEND _CXX_FLAGS ${BFFLAGS_ASAN})
-        endif()
-        if(ENABLE_USAN)
-            list(APPEND _C_FLAGS ${BFFLAGS_USAN})
-            list(APPEND _CXX_FLAGS ${BFFLAGS_USAN})
-        endif()
-        if(ENABLE_CODECOV)
-            list(APPEND _C_FLAGS ${BFFLAGS_CODECOV})
-            list(APPEND _CXX_FLAGS ${BFFLAGS_CODECOV})
-        endif()
     elseif(PREFIX STREQUAL "test")
         list(APPEND _C_FLAGS ${BFFLAGS_TEST} ${BFFLAGS_TEST_C} ${C_FLAGS_TEST})
         list(APPEND _CXX_FLAGS ${BFFLAGS_TEST} ${BFFLAGS_TEST_CXX} ${CXX_FLAGS_TEST})
